requirements.txt
# ABM Content Marketing Engine - Python Dependencies
# Version: 08-09-2025

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
supabase==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# AI/ML
anthropic==0.3.11

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
